[REFACTOR/VG-316] 툴팁 시스템 코드 정리 및 캐릭터 툴팁 연동 개선

+++ Source/GA6thFinal_Framework/GameScripts/Scripts/UI/Contents/TooltipDescriptionPanel.cpp
@@ -20,7 +20,7 @@ void TooltipDescriptionPanel::SetTooltips(const std::span<const int> ids)
 {
     ClearTooltips();
 
-    for (size_t i = 0; i < ids.size(); ++i)
+    for (int i = 0; i < ids.size(); ++i)
     {
         if (i < transform->GetChildCount())
         {
